+++ interface/src/lib/components/grow/marketMaking/createNew/amount/amountText.svelte
@@ -5,7 +5,7 @@
   export let tradingPair = "";
 </script>
 
-<div class="flex flex-col gap-2 mx-6">
+<div class="flex flex-col gap-2 mx-6 items-center">
   <span class="text-xl font-semibold">
     {$_("enter_deposit_amount")}
   </span>