+++ interface/src/lib/components/grow/marketMaking/createNew/confirmation/confirmHeader.svelte
@@ -4,13 +4,13 @@
   export let description: string;
 </script>
 
-<div class="space-y-2">
+<div class="space-y-2 text-start flex flex-col items-center">
   {#if label}
     <p class="text-xs font-semibold uppercase tracking-wide opacity-60">
       {label}
     </p>
   {/if}
-  <div class="flex flex-col gap-2">
+  <div class="flex flex-col gap-2 items-start justify-center">
     <span class="text-2xl font-bold text-base-content">
       {title}
     </span>
@@ -19,4 +19,3 @@
     </span>
   </div>
 </div>
-